fix(common): rethinking the gateway
