+++ aether-common-library/aether/common/keycloak/middleware.py
@@ -16,9 +16,67 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import jwt
+
+from django.conf import settings
+from django.contrib.auth import login, logout
 from django.contrib.auth.middleware import AuthenticationMiddleware
+from django.contrib.sessions.middleware import SessionMiddleware
+
+from aether.common.utils import find_in_request_headers, request as exec_request
+from aether.common.auth.utils import get_or_create_user_from_userinfo
 
 from .utils import check_user_token
+
+
+class GatewayAuthenticationMiddleware(SessionMiddleware):
+
+    def _logout(self, request):
+        request.session[settings.REALM_COOKIE] = None
+        request.session[settings.GATEWAY_HEADER_TOKEN] = None
+        logout(request)
+
+    def process_request(self, request):
+        # SessionMiddleware sets the session object in the request
+        # being accessible at `request.session`
+        super(GatewayAuthenticationMiddleware, self).process_request(request)
+
+        token = find_in_request_headers(request, settings.GATEWAY_HEADER_TOKEN)
+        if token:
+            try:
+                tokeninfo = jwt.decode(token, verify=False)
+                iss_url = tokeninfo['iss']
+
+                # check that the token url points to the keycloak server
+                if not iss_url.startswith(settings.KEYCLOAK_SERVER_URL):
+                    return
+
+                # go to ISS to get the realm+user info
+                realminfo = exec_request(method='get', url=iss_url).json()
+                realm = realminfo['realm']
+                userinfo_url = realminfo['token-service'] + '/userinfo'
+
+                # if this call fails the token is not longer valid
+                userinfo = exec_request(
+                    method='get',
+                    url=userinfo_url,
+                    headers={'Authorization': '{} {}'.format(tokeninfo['typ'], token)},
+                ).json()
+
+                request.session[settings.REALM_COOKIE] = realm
+                user = get_or_create_user_from_userinfo(request, userinfo)
+                login(request, user)
+
+                # flags that we are using the gateway to authenticate
+                request.session[settings.GATEWAY_HEADER_TOKEN] = realm
+
+            except Exception:
+                # something went wrong
+                self._logout(request)
+
+        elif request.session.get(settings.GATEWAY_HEADER_TOKEN):
+            # this session was using the gateway to authenticate before
+            self._logout(request)
 
 
 class KeycloakAuthenticationMiddleware(AuthenticationMiddleware):
